+++ src/services/DiffOpen.ts
@@ -1,9 +1,12 @@
 //	Imports ____________________________________________________________________
 
+import { spawn } from 'child_process';
 import * as fs from 'fs';
+import * as path from 'path';
 import * as vscode from 'vscode';
 
 import { Diff, File } from '../types';
+import { isMacOs, isWindows } from './common';
 import { DiffMessage } from './DiffMessage';
 
 //	Variables __________________________________________________________________
@@ -24,6 +27,8 @@ export class DiffOpen {
 		
 		msg.on('open:diffToSide', (data) => this.open(data, true));
 		msg.on('open:diff', (data) => this.open(data, vscode.workspace.getConfiguration('l13Diff').get('openToSide', false)));
+		
+		msg.on('reveal:file', (data) => this.reveal(data));
 		
 	}
 	
@@ -105,7 +110,55 @@ export class DiffOpen {
 		
 	}
 	
+	private reveal (data:any) :void {
+		
+		const pathname:string = data.pathname;
+		
+		if (isMacOs) showFileInFinder(pathname);
+		else if (isMacOs) showFileInFinder(pathname);
+		else showFileInFolder(pathname);
+		
+	}
+	
 }
 
 //	Functions __________________________________________________________________
 
+function showFileInFinder (pathname:string) {
+	
+	const process = spawn('open', ['-R', pathname || '/']);
+	
+	process.on('error', (error:Error) => {
+		
+		process.kill();
+		vscode.window.showErrorMessage(error.message);
+		
+	});
+	
+}
+
+function showFileInExplorer (pathname:string) {
+	
+	const process = spawn('explorer', ['/select,', pathname || 'c:\\']);
+	
+	process.on('error', (error:Error) => {
+		
+		process.kill();
+		vscode.window.showErrorMessage(error.message);
+		
+	});
+	
+}
+
+function showFileInFolder (pathname:string) {
+	
+	const process = spawn('xdg-open', [path.dirname(pathname) || '/']);
+	
+	process.on('error', (error:Error) => {
+		
+		process.kill();
+		vscode.window.showErrorMessage(error.message);
+		
+	});
+	
+}